Add images directory and add content to steak.html
I added a directory (images) where I'll store any image files to be
used on the site. I also added an image to the directory (steak.jpg).

I added content to the steak.html recipe page. The page now
contains an image, a list of ingredients, and steps to make the dish.